backend/requirements.txt
fastapi
uvicorn
transformers
torch
newspaper3k
beautifulsoup4